Editable expire dates and amounts

+++ app/bioshare/lib/utils/session_expired.dart
@@ -5,13 +5,16 @@ import '../main.dart';
 import './show_popup.dart';
 import '../models/view_model.dart';
 
-sessionExpired(BuildContext context) async {
+sessionExpired(BuildContext? context) async {
+  context ??= App.navigatorKey.currentContext;
+  if (context == null) return;
+
   Navigator.of(context).pop();
-  showPopup(context, "Twoja sesja wygasła", "musisz się zalogować ponownie");
   await App.secureStorage.delete(key: "jwt");
 
   if (context.mounted) {
     Navigator.of(context).popUntil((route) => route.isFirst);
     Provider.of<ViewModel>(context, listen: false).goToLogin();
+    showPopup(context, "Twoja sesja wygasła", "musisz się zalogować ponownie");
   }
 }
